Add validators for all user methods

+++ app/api/controllers/user.js
@@ -6,6 +6,7 @@ var userService        = require('../services/user'),
 	validators         = require('../validators/user');
 
 
+//Getting list of all users
 module.exports.getAllUsers = function(req, res) {
 	userService.getAllUsers()
 		.then(function(users) {
@@ -18,7 +19,17 @@ module.exports.getAllUsers = function(req, res) {
 };
 
 
+//Getting user by user id
 module.exports.getUser = function(req, res) {
+	//Validation of user id
+	if (!validators.validateUserId(req.params.userId)) {
+		var err = `User id '${req.params.userId}' is not valid`;
+		console.log(err);
+		res.json(presenter.fail(null, err));
+		return;
+	}
+
+	//Fetch user if user id is valid
 	userService.getUser(req.params.userId)
 		.then(function(user) {
 			res.json(presenter.success(user));
@@ -30,24 +41,47 @@ module.exports.getUser = function(req, res) {
 };
 
 
+//Creation new user
 module.exports.createUser = function(req, res) {
-	var errors = validators.createUser(req.body);
-	if (errors) {
-		res.json(presenter.fail(null, errors));
-	} else {
-		userService.createUser(req.body)
-			.then(function(user) {
-				res.json(presenter.success(user));
-			})
-			.catch(function(err) {
-				console.log(err);
-				res.json(presenter.fail(err, 'Error occurred while creating user'));
-			});
+	//Validation of user data like
+	//login, password, firstname, lastname
+	var err = validators.createUser(req.body);
+	if (err) {
+		console.log('User creation error:', err);
+		res.json(presenter.fail(null, err));
+		return;
 	}
+
+	//Create user if data are valid
+	userService.createUser(req.body)
+		.then(function(user) {
+			res.json(presenter.success(user));
+		})
+		.catch(function(err) {
+			console.log(err);
+			res.json(presenter.fail(err, 'Error occurred while creating user'));
+		});
 };
 
 
+//Create following for user
 module.exports.createFollower = function(req, res) {
+	//Validate user id and following user id
+	var err;
+	if (!validators.validateUserId(req.params.userId)) {
+		err.userId = 'User id is not valid';
+	}
+	if (!validators.validateUserId(req.params.followingUserId)) {
+		err.followingUserId = 'Following user id is not valid';
+	}
+
+	if (err) {
+		console.log('Error of follower creation.', err);
+		res.json(presenter.fail(null, err));
+		return;
+	}
+
+	//Create link between users if ids are valid
 	userService.createFollower(req.params.userId, req.params.followingUserId)
 		.then(function() {
 			res.json(presenter.success(null));
@@ -59,7 +93,24 @@ module.exports.createFollower = function(req, res) {
 };
 
 
+//Removing following for user
 module.exports.removeFollower = function(req, res) {
+	//Validate user id and following user id
+	var err;
+	if (!validators.validateUserId(req.params.userId)) {
+		err.userId = 'User id is not valid';
+	}
+	if (!validators.validateUserId(req.params.followingUserId)) {
+		err.followingUserId = 'Following user id is not valid';
+	}
+
+	if (err) {
+		console.log('Error of follower remove.', err);
+		res.json(presenter.fail(null, err));
+		return;
+	}
+
+	//Remove link between users if ids are valid
 	userService.removeFollower(req.params.userId, req.params.followingUserId)
 		.then(function() {
 			res.json(presenter.success(null));
@@ -71,7 +122,17 @@ module.exports.removeFollower = function(req, res) {
 };
 
 
+//Deleting user
 module.exports.deleteUser = function(req, res) {
+	//Validate user id
+	if (!validators.validateUserId(req.params.userId)) {
+		var err = `User id '${req.params.userId}' is not valid`;
+		console.log(err);
+		res.json(presenter.fail(null, err));
+		return;
+	}
+
+	//Delete user if his id is valid
 	userService.removeUser(req.user.id, req.params.userId)
 		.then(function() {
 			res.json(presenter.success(null));
